+++ exprv/src/main/java/com/prenotazioni/exprivia/exprv/repository/UserRepository.java
@@ -1,15 +1,11 @@
 package com.prenotazioni.exprivia.exprv.repository;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.prenotazioni.exprivia.exprv.entity.Users;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Integer> {
-
-    
+public interface UserRepository extends JpaRepository<Users, Long> {
 
 }
